requirements-test.txt
pytest==3.0.5
pytest-flake8==0.8.1
pytest-aiohttp==0.1.3
